Moved nibble related stuff to it's own module
- It is big enough that it probably should be its own module.
- Upcoming changes are likely going to add more to `partial_trie.rs`.
